# Optimized image loading to reduce memory footprint and boot time