Preliminary testing (1.4.x branch) suggests that these changes reduce hardware memory requirements as follows:

Platform | RAM requirement
-- | --
Baremetal | 768MB ~~1280MB~~  (1024MB to run installer)
VirtualBox | 768MB ~~1280MB~~ (1024MB to run installer)
VMWare | 768MB ~~1280MB~~ (installed to disk) 1024MB ~~2048MB~~ (docker-machine) 

Seems to boot much faster:

```
RancherOS Boot Time*      v1.4 -> dev build
==================================================
VMware Fusion  2GB 2vPU      87sec -> 35sec
==================================================
VirtualBox     2GB 2vCPU     48sec -> 31sec
==================================================
VirtualBox     3GB 2vCPU     49sec -> 27sec
==================================================

*time from kernel starting to "RancherOS started" log message
```

Caveats? ISO file size increases approx. by 30MB.

![screen shot 2018-09-10 at 04 14 10](https://user-images.githubusercontent.com/3813921/45297731-9b049480-b506-11e8-9b4e-7f5dffc27e90.png)