@niusmallnan 2018-09-10
Thanks, this is great. @janeczku 
LGTM.

Related: https://github.com/rancher/os/issues/2392